requirement.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
